Splits the public and internal express apps, inserts worker queue between
